requirements.txt
bcrypt==4.0.1
boto3==1.26.80
botocore==1.29.80
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
folium==0.14.0
google-api-core==2.11.0
google-auth==2.16.1
google-cloud==0.34.0
google-cloud-vision==3.4.0
googleapis-common-protos==1.58.0
googlemaps==4.10.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
numpy==1.24.2
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
